requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
bcrypt==3.2.0
billiard==4.2.1
boto3==1.38.35
botocore==1.38.35
cachetools==5.5.2
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.15.0
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fasta2a==0.2.16
fastapi==0.115.2
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.20.0
greenlet==3.1.1
griffe==1.7.3
groq==0.27.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
logfire-api==3.18.0
markdown-it-py==3.0.0
mcp==1.9.3
mdurl==0.1.2
mistralai==1.8.2
motor==3.6.0
numpy==2.2.5
openai==1.86.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-ai==0.2.16
pydantic-ai-slim==0.2.16
pydantic-evals==0.2.16
pydantic-graph==0.2.16
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.6
starlette==0.40.0
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
zstandard==0.23.0
reportlab==4.2.0
python-docx==1.1.0
beautifulsoup4==4.12.3
